Исправил ошибки..

+++ src/BankAccount.java
@@ -11,7 +11,7 @@ class BankAccount {
     public void deposit(double sum){
         amount += sum;
     }
-    public void withDraw(double sum) throws LimitException{
+    public void withDraw(int sum) throws LimitException{
         if (sum > amount){
             throw new LimitException("Недостаточно средств на счете", amount);
         }
